Split vegafusion-sql tests into separate files
