+++ vegafusion-sql/tests/test_stack.rs
@@ -1,0 +1,173 @@
+#[macro_use]
+extern crate lazy_static;
+
+mod utils;
+use utils::{TOKIO_RUNTIME, make_connection, check_dataframe_query};
+use datafusion_expr::{col, expr, lit, round, Expr};
+use rstest::rstest;
+use serde_json::json;
+use std::ops::{Div, Mul};
+use std::sync::Arc;
+use vegafusion_common::{
+    error::Result,
+    data::table::VegaFusionTable
+};
+use vegafusion_dataframe::dataframe::{DataFrame, StackMode};
+use vegafusion_sql::connection::SqlConnection;
+use vegafusion_sql::dataframe::SqlDataFrame;
+
+
+fn stack_data(conn: Arc<dyn SqlConnection>) -> Arc<dyn DataFrame> {
+    let table = VegaFusionTable::from_json(
+        &json!([
+                {"a": 1, "b": 9, "c": "A"},
+                {"a": -3, "b": 8, "c": "BB"},
+                {"a": 5, "b": 7, "c": "A"},
+                {"a": -7, "b": 6, "c": "BB"},
+                {"a": 9, "b": 5, "c": "BB"},
+                {"a": -11, "b": 4, "c": "A"},
+                {"a": 13, "b": 3, "c": "BB"},
+            ]),
+        1024,
+    )
+        .unwrap();
+
+    SqlDataFrame::from_values(&table, conn).unwrap()
+}
+
+fn make_stack_for_mode(df: Arc<dyn DataFrame>, mode: StackMode) -> Result<Arc<dyn DataFrame>> {
+    df.stack(
+        "a",
+        vec![Expr::Sort(expr::Sort {
+            expr: Box::new(col("b")),
+            asc: true,
+            nulls_first: true,
+        })],
+        &["c".to_string()],
+        "start",
+        "end",
+        mode,
+    )
+        .and_then(|df| {
+            df.sort(
+                vec![
+                    Expr::Sort(expr::Sort {
+                        expr: Box::new(col("c")),
+                        asc: true,
+                        nulls_first: true,
+                    }),
+                    Expr::Sort(expr::Sort {
+                        expr: Box::new(col("end")),
+                        asc: true,
+                        nulls_first: true,
+                    }),
+                ],
+                None,
+            )
+        })
+}
+
+#[cfg(test)]
+mod test_mode_zero {
+    use crate::*;
+
+    #[rstest(
+        dialect_name,
+        case("athena"),
+        case("bigquery"),
+        case("clickhouse"),
+        case("databricks"),
+        case("datafusion"),
+        case("dremio"),
+        case("duckdb"),
+        case("mysql"),
+        case("postgres"),
+        case("redshift"),
+        case("snowflake"),
+        case("sqlite")
+    )]
+    fn test(dialect_name: &str) {
+        println!("{dialect_name}");
+        let (conn, evaluable) = TOKIO_RUNTIME.block_on(make_connection(dialect_name));
+
+        let df = stack_data(conn);
+        let df_result = make_stack_for_mode(df, StackMode::Zero);
+        check_dataframe_query(df_result, "stack", "mode_zero", dialect_name, evaluable);
+    }
+}
+
+#[cfg(test)]
+mod test_mode_center {
+    use crate::*;
+
+    #[rstest(
+        dialect_name,
+        case("athena"),
+        case("bigquery"),
+        case("clickhouse"),
+        case("databricks"),
+        case("datafusion"),
+        case("dremio"),
+        case("duckdb"),
+        case("mysql"),
+        case("postgres"),
+        case("redshift"),
+        case("snowflake"),
+        case("sqlite")
+    )]
+    fn test(dialect_name: &str) {
+        println!("{dialect_name}");
+        let (conn, evaluable) = TOKIO_RUNTIME.block_on(make_connection(dialect_name));
+        let df = stack_data(conn);
+        let df_result = make_stack_for_mode(df, StackMode::Center);
+        check_dataframe_query(df_result, "stack", "mode_center", dialect_name, evaluable);
+    }
+}
+#[cfg(test)]
+mod test_mode_normalized {
+    use crate::*;
+
+    #[rstest(
+        dialect_name,
+        case("athena"),
+        case("bigquery"),
+        case("clickhouse"),
+        case("databricks"),
+        case("datafusion"),
+        case("dremio"),
+        case("duckdb"),
+        case("mysql"),
+        case("postgres"),
+        case("redshift"),
+        case("snowflake"),
+        case("sqlite")
+    )]
+    fn test(dialect_name: &str) {
+        println!("{dialect_name}");
+        let (conn, evaluable) = TOKIO_RUNTIME.block_on(make_connection(dialect_name));
+        let df = stack_data(conn);
+        let df_result = make_stack_for_mode(df, StackMode::Normalize);
+        // Round start and end to 2 decimal places to avoid numerical precision issues when comparing results
+        let df_result = df_result.and_then(|df| {
+            df.select(vec![
+                col("a"),
+                col("b"),
+                col("c"),
+                round(col("start").mul(lit(100)))
+                    .div(lit(100))
+                    .alias("trunc_start"),
+                round(col("end").mul(lit(100)))
+                    .div(lit(100))
+                    .alias("trunc_end"),
+            ])
+        });
+        check_dataframe_query(
+            df_result,
+            "stack",
+            "mode_normalized",
+            dialect_name,
+            evaluable,
+        );
+    }
+}
+
